django_backend/requirements.txt
asgiref==3.8.1
boto3==1.35.86
botocore==1.35.86
certifi==2024.12.14
charset-normalizer==3.4.0
Django==4.2.17
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
jmespath==1.0.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
stripe==11.4.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.15
